refactored functions into planningUtils
